MS-79: refactor inputs for profiles and sessions forms
