+++ src/Components/Sessions/Form/index.js
@@ -2,9 +2,9 @@ import { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import useQuery from '../../../Hooks/useQuery';
 import Select from '../Select';
-import TextArea from '../TextArea';
-import Input from '../Input';
+import Textarea from '../../Shared/textarea';
 import styles from './form.module.css';
+import Input from '../../Shared/input';
 
 function sessionsForm() {
   const [dateValue, setDateValue] = useState('');
@@ -187,6 +187,7 @@ function sessionsForm() {
           required
         />
         <Input
+          title="Date"
           name="date"
           value={dateValue}
           onChange={onChangeDateInput}
@@ -194,7 +195,8 @@ function sessionsForm() {
           disabled={isLoading}
           required
         />
-        <TextArea
+        <Textarea
+          title="Notes"
           name="notes"
           value={notesValue}
           onChange={onChangeNotesInput}
