--- src/Components/Sessions/Input/input.module.css
@@ -1,8 +1,0 @@
-.input {
-  height: 56px;
-  margin: 10px;
-  padding: 0 10px;
-  border: 1px solid gray;
-  border-radius: 4px;
-  font-size: 16px;
-}
